requirements.txt
apache-airflow-providers-postgres==5.4.0
pytest == 7.2.0
testcontainers == 3.7.0
